Refactor postprocessing and props interface
Use PropError for handling errors in computing props. This is far from
ideal, but the semantics are close enough to correct for now, and it
avoids having to construct a result in the case of infallible props.

Accept a flat, key-first list of props. This is a compromise between
clarity and usability. It prioritizes per-key props over supporting
TermHandlers, on the grounds that TermHandlers will soon themselves be
refactored.