Parse eof as closing strings
Instead of handling unclosed strings with repairDelimiters. This
makes it more difficult to change the repair method in the future,
but fixes the failing test, in which a backslash added by delimiter
repair was parsed into the `raw` field of a text content node.

Alternative solution: extend `trimCst()` to support truncating the raw
field as necessary.